+++ examples/handshake/server.py
@@ -8,7 +8,7 @@ secret_code = "pancakes"
 
 class CustomDaemon(Pyro4.Daemon):
     def validate_handshake(self, conn, data):
-        print("Daemon received handshake reqest from:", conn.sock.getpeername())
+        print("Daemon received handshake request from:", conn.sock.getpeername())
         print("Handshake data:", data)
         # if needed, you can inspect Pyro4.current_context
         if data == secret_code:
@@ -18,6 +18,9 @@ class CustomDaemon(Pyro4.Daemon):
         else:
             print("Secret code wrong! Connection refused.")
             raise ValueError("wrong secret code, connection refused")
+
+    def client_disconnect(self, conn):
+        print("Daemon client disconnects:", conn.sock.getpeername())
 
 
 daemon = CustomDaemon()