+++ docs/source/changelog.rst
@@ -15,6 +15,8 @@ Change Log
   to customize this and add a handshake/connection validator in the daemon.
 - drastically reduced the overhead of creating a new proxy connection by piggybacking the metadata on the
   connection response (this avoids a separate remote call to get_metadata). New proxy connections are ~50% faster.
+- added ``Daemon.client_disconnect()`` as a hook for when clients disconnect (``Daemon.validate_handshake`` can
+  be used as the hook to handle new connections)
 - added handshake example
 - fixed callcontext example daemon to actually return a custom annotation
 - fixed benchmark/connections example